// src/modules/invoice/enum/Invoice.enum.ts
export enum ItemStatus {
  SUCCESS = 'success',
  FAILED = 'failed', 
}

export enum InvoiceType {
  PURCHASE = 'purchase',   
  SALE = 'sale',           
  STOCK = 'stock',         
  DAMAGED = 'damaged',     
}

export enum InvoiceStatus {
  SUCCESS = 'success',
  PARTIAL_SUCCESS = 'partial_success',
}